Support detecting perl files without shebang line
This enables getting a IContentDescription for pl or pm files that do not
have a shebang line. Even though such descriptions are currently not used
by Eclipse itself for opening editors a third-party may have the need for
going through the content type for finding an editor to open perl files.