Introduced org.epic.perleditor.requiredFileResolver extension point
This patch introduces a new extension point which permits expressing
custom algorithms for locating files specified with the 'require'
statement. Such handlers will be considered first by the 'Open
Declaration' functionality which traverses the set of included files
when trying to look for a declaration.

This allows implementing custom logic for expressions like

  require customGlobalObject->myLibraryFile;

or similar: users can manually evaluate the variable name and return an
appropriate File object (or null, if the given string cannot be resolved
to any File).