Add Eclipse Public License to project
I was considering installing this plugin, and found myself overlooking the health of the project.

Good going on moving the codebase over to GitHub!

One of the suggested "community health" checks for the project was to include a license file.  It really looks like the [license file over on the Eclipse Perl Integration home page](http://epic-ide.org/license.php) is most similar to the Eclipse Public License file, so I thought I'd propose it.

Cheers!